forces support file with checking routines
